+++ src/search.ts
@@ -5,12 +5,15 @@ import filterColorIdentity from "./search-filters/color-identity";
 import filterComboData from "./search-filters/combo-data";
 import filterSize from "./search-filters/size";
 import filterTags from "./search-filters/tags";
+import sortCombos from "./sort-combos";
 import createMessage from "./parse-query/create-message";
 
 import type { SearchResults } from "./types";
 
 export default async function search(query = ""): Promise<SearchResults> {
   const searchParams = parseQuery(query);
+  const sort = searchParams.sort || "colors";
+  const order = searchParams.order || "ascending";
   const { errors } = searchParams;
 
   let combos = await lookupApi();
@@ -20,6 +23,7 @@ export default async function search(query = ""): Promise<SearchResults> {
   combos = filterComboData(combos, searchParams);
   combos = filterSize(combos, searchParams);
   combos = filterTags(combos, searchParams);
+  combos = sortCombos(combos, sort, order);
 
   return {
     errors,